+++ src/index.ts
@@ -1,2 +1,4 @@
+export * from './types'
+export * from './utils'
 export * from './prismy'
 export * from './testHandler'
